Patch for ARM 32 build.
Antl had a problem with ARM 32:
https://github.com/antlr/antlr4/issues/1858

So picking the commit of antlr where the PR #1990 was merged.

Also casting a constant to uint64 (see default type of an untyped constant
https://golang.org/ref/spec#Constants ).

+++ go/lib/addr/isdas.go
@@ -160,7 +160,7 @@ func (as AS) inRange() bool {
 
 func (as AS) MarshalText() ([]byte, error) {
 	if !as.inRange() {
-		return nil, common.NewBasicError("invalid AS", nil, "max", MaxAS, "actual", as)
+		return nil, common.NewBasicError("invalid AS", nil, "max", uint64(MaxAS), "actual", as)
 	}
 	return []byte(as.String()), nil
 }
